Add Docker setup and enhance CI/CD pipeline
Added Docker support and enhancements in CI/CD pipeline. Created Dockerfile, docker-compose.yml and .dockerignore to enable containerization of the application. Updated Makefile to include new commands for building and testing the application. Renamed CI workflow from 'ci.yaml' to 'ci.yml' and added steps to build docker image and push it to DockerHub on successful lint and test run in the GitHub Actions. Added '/bin/' in .gitignore to ignore the binaries built during Docker build.
The changes are essential to streamline and automate the process of application deployment and ensure code quality before merges into mainline branch.