Update CI/CD workflow and remove docker build step
Removed the Docker build step from the GitHub Actions workflow file 'ci.yml'. This change was made because the Docker build and push action now directly depends on the successful execution of linting and testing. By making this modification, we have streamlined the continuous integration and deployment process, thus saving time and resources.