Added password hashing and comparison to account operations
Password hashing and comparison functionality was introduced to the account operations in MongoDB and PostgreSQL databases. PasswordHelper is used in MongoAccountDatabase for MongoDB and AccountPostgresDatabase for Postgres. This ensures password security during storing and retrieval. The password is hashed before it's saved into the database during the account registration process. When authenticating a user, the entered password is compared with the hashed password stored in the database. If there's a mismatch, it returns an error message 'invalid credentials'.