Create auth server script, refactor adapters into packages, improve imports
This commit adds the script `auth_server.go` in `scripts` package to reduce complexity in starting server from `main.go`. The import redundancy in the database files of `account` and `token` packages in `adapters` are cleaned. The `adapters` folder is refactored into `pkg` for a more comprehensive architecture, facilitating the import process in different modules. Also, there is a correction in adaptors import usages in `main.go`. The changes improve the usability of the packages and enhances the entire application's structure keeping the design pattern of the Go language in mind.