Remove auth_server script, improve HTTP server setup and auth handling
This commit removes the auth_server.go script and improves the HTTP server setup and authentication handling. The changes were made to enhance the organization of the code and ease of use. New API response models were introduced with AuthenticationRequest, AuthenticationResponse, LogoutRequest, APIResponse and APIError structured. Routes for authentication were added in the auth.go in new 'routes' package. Auth handling was moved to separate handlers in newly created 'handlers' package.

Finally, the server.go script was introduced to start the HTTP server. It receives dependency injection for account and token repositories and sets up auth routes. The issue with the deletion of the token in the redis.go adapter was also addressed.