Refactor script path in docker build-push action
The path to the Docker build context has been adjusted in the cd.yml workflow file for the build-push-action in GitHub Actions. The context path was previously 'src/scripts', but has been updated to './src/scripts'. This change ensures we maintain the integrity and reliability of the docker build process in our project's CD pipeline.