Update docker context in GitHub CD workflow
This commit primarily changes the docker build context and the Dockerfile's path directly in our continuous deployment pipeline on GitHub. The context was moved from 'src/scripts' to the root directory '.', and the Dockerfile's route has been adjusted to './src/scripts/Dockerfile'. This alteration guarantees a more robust and reliable docker build process which ultimately leads to a smoother and more efficient CD pipeline.